Enable co19 test suite on dart test scripts.
Updated the TestExpectations class to be able to deal with the
odd matching rules that are used for co19.

Currently we are using a lot of memory when running all of these
tests. I have optimized the test expectation part quite a bit.
The thing that seems bad now if the processing of all the lines
of every simple test to find DartOptions. We need to work on that.

R=whesse@google.com
BUG=
TEST=

Review URL: http://codereview.chromium.org//8539044

git-svn-id: http://dart.googlecode.com/svn/branches/bleeding_edge/dart@1563 260f80e4-7a28-3924-810f-c04153c831b5