Test for finding correct method for overriden native method of hidden native class.
I think the current implementation of $patchMethod has bugs that make it 'mostly work'.  It patches one step too high (either the object or the prototype of the subclass).  That is why I needed to add test#2 to exhibit the behaviour I thought I would see on the first test.

BUG=
TEST=

Review URL: http://codereview.chromium.org//8674009

git-svn-id: http://dart.googlecode.com/svn/branches/bleeding_edge/dart@1815 260f80e4-7a28-3924-810f-c04153c831b5